+++ packages/riverpod_analyzer_utils/lib/src/nodes/annotation.dart
@@ -47,8 +47,10 @@ extension RiverpodAnnotatedAnnotatedNodeX on Annotation {
         (e) => e.expression.providerDependencyList,
       );
 
-      final AstNode? retryNode = arguments?.named('retry')?.expression;
-      if (retryNode is! SimpleIdentifier?) {
+      final retryNode = arguments?.named('retry')?.expression;
+      final parsedRetryNode = retryNode.let(ConstantSymbol.tryParse);
+
+      if (retryNode != null && parsedRetryNode == null) {
         errorReporter(
           RiverpodAnalysisError.ast(
             'The "retry" argument must be a variable.',
@@ -64,7 +66,7 @@ extension RiverpodAnnotatedAnnotatedNodeX on Annotation {
         keepAliveNode: arguments?.named('keepAlive'),
         dependenciesNode: dependenciesNode,
         dependencyList: dependencyList,
-        retryNode: retryNode as SimpleIdentifier?,
+        retryNode: parsedRetryNode,
       );
     });
   }
@@ -85,7 +87,39 @@ final class RiverpodAnnotation {
   final NamedExpression? keepAliveNode;
   final NamedExpression? dependenciesNode;
   final ProviderDependencyList? dependencyList;
-  final SimpleIdentifier? retryNode;
+  final ConstantSymbol? retryNode;
+}
+
+sealed class ConstantSymbol {
+  const ConstantSymbol();
+
+  static ConstantSymbol? tryParse(AstNode node) {
+    // Use a switch with type patterns instead of if-else chains
+    switch (node) {
+      case PrefixedIdentifier():
+        return PrefixedIdentifierConstantSymbol(node);
+      case SimpleIdentifier():
+        return SimpleIdentifierConstantSymbol(node);
+      default:
+        return null;
+    }
+  }
+
+  AstNode get node;
+}
+
+final class PrefixedIdentifierConstantSymbol extends ConstantSymbol {
+  PrefixedIdentifierConstantSymbol(this.node);
+
+  @override
+  final PrefixedIdentifier node;
+}
+
+final class SimpleIdentifierConstantSymbol extends ConstantSymbol {
+  SimpleIdentifierConstantSymbol(this.node);
+
+  @override
+  final SimpleIdentifier node;
 }
 
 final class RiverpodAnnotationElement {