+++ packages/riverpod_generator/test/integration/hash/retry.dart
@@ -10,3 +10,6 @@ Duration? myRetry2(int retryCount, Object error) => null;
 
 @Riverpod(retry: myRetry2)
 String b(Ref ref, int arg) => throw UnimplementedError();
+
+@Riverpod(retry: ProviderContainer.defaultRetry)
+String c(Ref ref) => throw UnimplementedError();