+++ packages/riverpod_generator/test/retry_test.dart
@@ -1,3 +1,4 @@
+import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:test/test.dart';
 
 import 'integration/hash/retry.dart';
@@ -8,5 +9,7 @@ void main() {
 
     expect(bProvider.retry, myRetry2);
     expect(bProvider(42).retry, myRetry2);
+
+    expect(cProvider.retry, ProviderContainer.defaultRetry);
   });
 }